Don't auto dump json
Motivation:

On adding swagger, the schema validation runs on what reply returns. If
reply returns the body as a string, the schema validation will fail.

Modifications:

Don't auto dump json.

Result:

This actually doesn't have an impact on the services, because there was
something on the pipeline that already converted to json.